Update 9
Renamed the base folder for the project.  Moved around experimental js and html files into a separate folder.  Created a display variables component to display the variables to the screen as they change.  Completed calcBayes method that calculates the answer probability using bayes formula.  Created testAnswer method that checks to make sure the answer is a number between 0 and 1 and if it is converts it to a decimal of 2 places.  Created a display message variable that updates the state answer with the calculated answer.  Added display components to main app.  Added some css grid styles to the css to begin work positioning the elements.